Fix panic issue for function mesh (#36)
Signed-off-by: xiaolong.ran <rxl@apache.org>

Currently, if we do not specify certain fields in the yaml file, it will cause the program to panic. This is because in the code, we have not added a check on whether the field is nil. This pr will handle these problems, increase the nil check, and set default values for them.