Support authorization for java proxy (#74)
Fixed #58 

## Motivation

At first, I thought I needed to define the authorization cache configuration class when obtaining authorization, for example, by implementing a class similar to `ConfigurationCacheService`. However, after my research, I found that the current the class `AuthenticationService` value receives class `ConfigurationCacheService` as an argument, and this class uses zookeeper as a caching service. java-proxy runs in the same place as the pulsar cluster and inherits the function_worker configuration, so it can also use zookeeper as a caching service.

For the authentication plugin, the custom logic is implemented in jetty's filter, so there is no need to update the