Use {{ .Release.Namespace }} to replace {{ .Values.installation.namespace }} in chart templates
When I deployed the function-mesh in kubernetes according to this [document](https://functionmesh.io/docs/install-function-mesh#install-function-mesh-through-helm),
I found that the operator was not created in the expected namespace ("function-mesh"),
but in a namespace called "function-mesh-system".

It's because in `values.yaml` we have set the following,
and also the reason for changing the namespace where the operator is created:
```yaml
installation:
  namespace:
```

I think we can replace `.Values.installation.name` with helm's built-in `.Release.Namespace`,
which will simplify and clarify the installation process.

And the `templates/controller-manager-ns.yaml` file can also be replaced by Helm `--create-namespace` argument.

Signed-off-by: laminar <tpiperatgod@gmail.com>